requirements.txt
arcade==2.5.5
asgiref==3.4.1
attrs==21.2.0
blocking==0.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.9
django-banjo==0.2.8
django-extensions==3.1.5
django-heroku==0.0.0
easing-functions==1.0.4
gunicorn==20.1.0
idna==3.3
invoke==1.6.0
numpy==1.20.1
pandas==1.3.4
Pillow==8.4.0
pycparser==2.20
PyGithub==1.55
pyglet==1.5.21
PyJWT==2.3.0
pymunk==6.0.0
PyNaCl==1.4.0
python-dateutil==2.8.2
pytiled-parser==0.9.4a3
pytz==2021.3
PyYAML==5.4.1
questgame==0.1.11
requests==2.26.0
Shapely==1.7.1
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.3
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.13.3
xvfbwrapper==0.2.9
